// types/kms.ts
export type UserRole = 'guest' | 'staff' | 'admin';

export interface UserAccount {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  status: 'Aktif' | 'Menunggu' | 'Nonaktif';
  department: string;
  initial: string;
}

export type PageId = 'index' | 'dashboard' | 'map' | 'modules' | 'knowledge' | 'chat' | 'admin' | 'baseline' | 'login';